fix: resolve open PR #101 review comments
- fix(LoadfileOnlyGenerator): remove (int) cast on FileCount; use long
  totalLines to prevent value truncation before ChaosEngine constructor
  guards run (addresses coderabbitai comment on src/ChaosEngine.cs:53)

- fix(E2E/bat): strengthen Test 5 chaos-filter assertion in
  run-e2e-loadfile.bat to parse JSON with PowerShell and verify
  totalAnomalies > 0 and injectedAnomalies is non-empty, and that only
  allowed types are present (addresses coderabbitai comment on
  tests/run-e2e-loadfile.bat:208)

- refactor(ChaosEngine): rename ApplyOptBatesIdCorruption to
  ApplyOptBatesNumberCorruption and update audit description from
  "Bates ID" to "Bates Number" to align with project ubiquitous
  language (addresses coderabbitai comment on src/ChaosEngine.cs:298)

- refactor(tests): update test method name OptBatesId_RemovesBatesId to
  OptBatesId_RemovesBatesNumber and update inline comments in
  LoadfileOnlyGeneratorTests.cs and run-e2e-loadfile.sh to use
  canonical term "Bates Number"

+++ src/LoadfileOnlyGenerator.cs
@@ -26,9 +26,9 @@ internal static class LoadfileOnlyGenerator
 
         var encoding = EncodingHelper.GetEncodingOrDefault(request.Encoding);
         var eolString = GetEolString(request.EndOfLine);
-        int totalLines = request.LoadFileFormat == LoadFileFormat.Opt
-            ? (int)request.FileCount // OPT has no header
-            : (int)request.FileCount + 1; // DAT: +1 for header
+        long totalLines = request.LoadFileFormat == LoadFileFormat.Opt
+            ? request.FileCount // OPT has no header
+            : request.FileCount + 1; // DAT: +1 for header
 
         // Initialize chaos engine if enabled
         ChaosEngine? chaosEngine = null;
@@ -207,7 +207,7 @@ internal static class LoadfileOnlyGenerator
         Random random)
     {
         // Opticon 7-column comma-separated format:
-        // BatesID,Volume,ImagePath,DocBreak(Y/blank),BoxBreak,FolderBreak,PageCount
+        // BatesNumber,Volume,ImagePath,DocBreak(Y/blank),BoxBreak,FolderBreak,PageCount
         var now = request.Seed.HasValue ? new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc) : DateTime.UtcNow;
         var buffer = new StringBuilder();
